backend/requirements.txt
# FastAPI e dependências
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
alembic==1.12.1

# Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP e APIs
requests==2.31.0
httpx==0.25.1
aiohttp==3.9.0

# Dados e processamento
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# Geolocalização
geopy==2.4.0
shapely==2.0.2

# Utilidades
python-dotenv==1.0.0
schedule==1.2.0
aiofiles==23.2.1

# Firebase (notificações)
firebase-admin==6.2.0

# Logging e monitoring
loguru==0.7.2

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
